[#24] rename: 재사용 가능 컴포넌트들 파일 위치 변경
