+++ src/pages/Quiz/AIQuiz/CreateAIQuiz/index.tsx
@@ -2,8 +2,8 @@ import { useSearchParams } from 'react-router-dom';
 import { useEffect, useRef, useState } from 'react';
 import { useSetRecoilState } from 'recoil';
 import RightSideBar from './RightSideBar';
-import UploadType from './UploadType';
-import TextType from './TextType';
+import UploadType from '../../../../components/SelectAIType/UploadType';
+import TextType from '../../../../components/SelectAIType/TextType';
 import { UploadedFileType } from '../../../../types';
 import loadingSelector from '../../../../recoil/selectors/loading';
 import SelectAIType from '../../../../components/SelectAIType';
